YomiChat/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
smolagents
smoltools-jinaai
duckduckgo-search
liteLLM
requests
